Fix invalid cfg attr & non_canonical_partial_ord
In the antiflood-tokens module, one of the #[cfg]s used an incorrect
attribute: features instead of feature. This inadvertently gated a few
lines of code.

I also fixed two small lints. PartialOrd should only be manually
implemented if Ord isn't implemented. Ord implies PartialOrd, so a
PartialOrd impl should forward to Ord.

Clippy denies ignoring the index for enumerate() - it's likely that the
caller intended to use it and forgot, so denying this lint is useful for
finding logic bugs. In this case, it doesn't seem like the index was
meant to be used anywhere.